Add very, very slow drag+drop.
We push all state changes through Replicache to gain access to its
conflict resolution. If another client is moving an object at the
same time, we should merge their movements, or something otherwise
reasonable should happen.

Replicache is not yet optimized for this type of usage - all writes
go to disk. So it's not very pretty right now, but it's correct.

Will fix Replicache subsequently.