Change the way registration of mutators is done with Replicache.
TypeScript can infer the shape of the Data object if we construct an
object rather than a class, which allows us to avoid typing the
signature twice and removes the temptation to abstract out the
signature types, removing tons of goop.

This in turn enables removing the mutators.ts file, which was really
just a place to hide all the ts goop. There's no need for the
distinction between this and the domain objects.

We could actually infer wrap this up more, and general wrapper
functions for our mutations that we feed right into rep.register(),
but I decided it made sense to be explicit here.

It's not that much work, and you should realize when you are
changing both these places.