Collapse mutations a different way.
When we drag an object, we get a pattern of mutations at server-side
like:

setCursor,moveShape,setCursor,moveShape,etc.

Because of this, our collapsing logic was failing to collapse because
it just looked for strings of consecutive mutations of same type.

We can't just collapse all events of the same type in the same push
because that could change the behavior of intervening mutations.

Instead, we use the same trick the client does - caching the accessed
data in memory.

This also had a side-effect of allowing us to parallelize writes to
MySQL (at the cost of having to keep the tx open for lenght of push).