Refactor: give up on svg layout.
Working with svg as a layout system is frustrating and buggy. It's
easier if you put each shape in its own svg container and move that
around using normal HTML primitives.

Also factored out a Rect shape which is stateless and a
RectController which has the smarts. I think this will make things
easier to follow when I add resize.