feat: sync descriptions and comments in reverse issue modified order (#97)
The default initial view is a list of issues in reverse modified order.  
This should inform the order in which we sync entries during
partial sync (what we call the incremental initial sync).
We return all issue meta data (what is displayed in list view) in the 
first pull response.
We then incrementally return descriptions and comments over a 
series of pull requests until partial sync is complete.
Prior to this change we synced these descriptions and comments
in 'key' order.
It is preferable to sync them in reverse modified order of their
corresponding issue, so that if a user clicks on an issue near the top 
of the default initial list view the entries needed for displaying the 
detail view is available as soon as possible.