overlay spaces rather than copying template space (#98)
### Problem
The current approach of giving each space its own copy of the template, while conceptually simple, has several drawbacks
1. each space takes up roughly 50MB of database space.  If the demo gets significant traffic 
2. Initing a space is slow, roughly 10 seconds, because of the size and indexes involved.  We created functions to preinit spaces from the console, but even there only 10 can be done at a time while pegging the DB for 100 seconds.

### Solution
Instead of a 'template' space we create a 'base' space.  Each other space contain only the deltas from the base space.
When we query space with `spaceID` we overlay the  space with `spaceID` on top of the 'base' space.  This makes the queries more complicated and potential harder to optimize, but the db will be much smaller so optimization will be less necessary.