fix: In Replicache 13, cookies must increase if the data has (#155)
changed. If the cookie stays the same, Replicache ignores the
pull response.

This gave me an idea though that is broader than Repliear: can
just introduce the concept of 'pull ID', yet another counter :).

This makes it a little more concrete than 'cvr version' and it
would also apply to most of the other strategies.